icons proof-of-concept, needs simplifying
