+++ src/plugins/dataflow/components/dataflow-program.tsx
@@ -618,7 +618,6 @@ export class DataflowProgram extends BaseComponent<IProps, IState> {
     const sharedProgramModel = this.props.tileContent.sharedProgramData as SharedProgramDataType;
     sharedProgramModel.setProgramNodes(nodes);
     const dataRateStr = ProgramDataRates.find((item) => item.val === Number(this.props.programDataRate))?.text ?? "";
-    console.log("|| about to setProgramSamplingRate", dataRateStr);
     sharedProgramModel.setProgramSamplingRate(dataRateStr);
   };
