+++ src/plugins/shared-assets/icons/icon-utilities.tsx
@@ -45,6 +45,8 @@ const icons: Record<string, FunctionComponent<SVGProps<SVGSVGElement>>> = {
   [kSignalKey]: SignalIcon
 };
 
+console.log("| icons: ", icons);
+
 export function iconUrl(id: string) {
   return `${kIconPrefix}://${id}`;
 }
