+++ src/plugins/simulator/components/simulator-tile.tsx
@@ -33,9 +33,6 @@ export const SimulatorTileComponent = observer(function SimulatorTileComponent({
 
   const component = content.simulationData.component;
 
-  console.log(
-    "|| SIM has sharedProgramData:", content.sharedProgramData
-  );
   return (
     <div className="simulator-content-container">
       <BasicEditableTileTitle />
