--- src/plugins/simulator/simulations/potentiometer-servo/assets/animated-wires.tsx
@@ -1,34 +1,0 @@
-import React from 'react';
-
-export const potToA1 = () => (
-  <>
-    <svg className="pot-to-a1 line" width="59px" height="20px" viewBox="0 0 59 20" version="1.1">
-      <title>path-4</title>
-      <g id="Page-1" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
-          <g id="Path-2-Copy-3" transform="translate(0.5, 1.6704)" strokeWidth="3">
-              <g id="path-4-link-line">
-                  <path
-                    d="M0,16 L4.5,16 C10.0228475,16 12.5,13.522847 12.5,8 C12.5,2.477152 14.9771525,0 20.5,0 L57.8783423,0"
-                    className="path-to-a1 line"
-                    transform="translate(28.9392, 8) scale(1, -1) translate(-28.9392, -8)">
-                  </path>
-              </g>
-          </g>
-      </g>
-    </svg>
-    <svg className="pot-to-a1 dashes" width="59px" height="20px" viewBox="0 0 59 20" version="1.1">
-      <title>path-4</title>
-      <g id="Page-1" stroke="none" strokeWidth="1" fill="none" fillRule="evenodd">
-          <g id="Path-2-Copy-3" transform="translate(0.5, 1.6704)" strokeWidth="3">
-              <g id="path-4-link-dashes">
-                  <path
-                    d="M0,16 L4.5,16 C10.0228475,16 12.5,13.522847 12.5,8 C12.5,2.477152 14.9771525,0 20.5,0 L57.8783423,0"
-                    className="path-to-a1 dashes"
-                    transform="translate(28.9392, 8) scale(1, -1) translate(-28.9392, -8)">
-                  </path>
-              </g>
-          </g>
-      </g>
-    </svg>
-  </>
-);
