+++ src/plugins/simulator/simulations/potentiometer-servo/potentiometer-servo.scss
@@ -276,3 +276,21 @@ $output-node-value: #f9ecab;
   top: 218px;
   right: 101px;
 }
+
+.node-icon {
+  width: 8px;
+  height: 8px;
+  background-size: 10px 10px;
+  margin-right: 2px;
+}
+
+.stopwatch {
+  height: 12px;
+  position: relative;
+  top: 1px;
+  left: -2px;
+}
+
+.mini-node.type-live.output svg {
+  transform: translate(-4px, 1px) scale(0.7)
+}
